dont use time.Duration which doesnt have Equal() method in proto Signed-off-by: Owen Diehl <ow.diehl@gmail.com>

+++ pkg/storage/stores/index/index.go
@@ -170,7 +170,7 @@ func (m MonitoredReaderWriter) GetShards(
 			// 2) when we resolve shards from the index + blooms
 			// NB(owen-d): since this is measured by the callee, it does not include time in queue,
 			// over the wire, etc.
-			shards.Statistics.Index.ShardsDuration = time.Since(start)
+			shards.Statistics.Index.ShardsDuration = int64(time.Since(start))
 		}
 
 		return err
